% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStringSpecies.R
\name{createStringSpecies}
\alias{createStringSpecies}
\title{Utility function (Internal, not exported) to convert argument to strings}
\usage{
createStringSpecies(itemName, chosenItem, convertToCharacter, numChars)
}
\arguments{
\item{itemName}{Character string. Variable name as it exists in the data base}

\item{chosenItem}{User input value of variable listed in itemName}

\item{convertTocharacter}{Boolean. Should we convert the \code{chosenItem} to a character string (This depends on how the variable is declared in the database)}

\item{numchars}{Numeric scalar. Number of characters to format the numeric \code{chosenItem} as.}
}
\value{
A charachter string
}
\description{
Deal explicity with NESPP3 and NESPP4 codes.
inputs to get_ functions are required as strings when passed as a sql statement
This function converts numeric inputs to character strings
}
\examples{
\dontrun{
createStringSpecies(itemName="area",area=503,convertToCharacter=TRUE,numChars=3)
createStringSpecies(itemName="species_itis",species,convertToCharacter=TRUE,numChars=6)

}

}
