% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_absence_presence.R
\name{get_absence_presence}
\alias{get_absence_presence}
\title{Extract absence-presence data from obdbs}
\usage{
get_absence_presence(channel, species = "all", year = 1994,
  sex = "all", area = "all")
}
\arguments{
\item{channel}{RODBC object (see \code{\link{connect_to_database}})}

\item{species}{Numeric Vector. a specific species code (NESPP3) or set of codes. Defaults to "all" species.
#' Numeric codes are converted to VARCHAR2(3 BYTE) when creating the sql statement. Character codes are short character strings.}

\item{year}{Numeric vector. Years to search over}

\item{sex}{Character string. Default = "all". options "M" (male), "F" (female), "U" (unsexed)}

\item{area}{Character vector. Statistical areas. Default = "all".}
}
\value{
A list is returned:

  \item{data}{containing the result of the executed \code{$sql} statement}

  \item{sql}{containing the sql call}

  \item{colNames}{a vector of the table's column names}
}
\description{
For a specified region of interest, extracts all hauls and indicates presence or absence for
species of interest. Sampled by observers on commercial fishing boats
This data is extracted from obspp
}
\section{Reference}{

Use the data dictionary (\url{http://nova.nefsc.noaa.gov/datadict/}) for field name explanations.
Note: species codes (nespp4) are stored in the database as VARCHAR2(4 BYTE)
}

\examples{
\dontrun{
# extracts info for american flounder (plaice) (124)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_length_age(channel,species=124)
get_length_age(channel,species="124")


}

}
\seealso{
\code{\link{connect_to_database}}
}
