% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obdbs.R
\docType{package}
\name{obdbs}
\alias{obdbs}
\alias{obdbs-package}
\title{svdbs: Extract info from Database}
\description{
Tools to aid in extraction of data from the obdbs database
}
\details{
\itemize{
\item Get_ functions to access, species codes, length-weight etc.
}

To learn more about using \code{obdbs}, start with the vignette: \code{browseVignettes(package="obdbs")} or click the index below
}
