% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lengths.R
\name{get_lengths}
\alias{get_lengths}
\title{Extract Length data from obdbs}
\usage{
get_lengths(channel, year = 1994, species = "all", sex = "all",
  marketCode = F)
}
\arguments{
\item{channel}{an RODBC object (see \code{\link{connect_to_database}})}

\item{species}{a specific species code or set of codes. Either numeric or character vector. Defaults to "all" species.
Numeric codes are converted to VARCHAR2(4 BYTE) when creating the sql statement. Character codes are short character strings.}

\item{sex}{character vector. Default = "all". options "M" (male), "F" (female), "U" (unsexed)}
}
\value{
A list is returned:

  \item{data}{containing the result of the executed \code{$sql} statement}

  \item{sql}{containing the sql call}

  \item{colNames}{a vector of the table's column names}
}
\description{
Extract a list of length for speices sampled by observers on commercial fishing boats
This data is extracted from oblen
}
\section{Reference}{

Use the data dictionary (\url{http://nova.nefsc.noaa.gov/datadict/}) for field name explanations.
Note: species codes (nespp4) are stored in the database as VARCHAR2(4 BYTE)
}

\examples{
\dontrun{
# extracts info for cod (73)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_length_age(channel,species=73)

# extracts info for cod ("COD")
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_length_age(channel,"cod") or
get_length_age(channel,"co") or
get_length_age(channel,"COD")

}

}
\seealso{
\code{\link{connect_to_database}}
}
